Test UI
pressing esc pulls up a button which prints a message to the console. the text has a different transform than the button and it makes me want to die because i have no good solution other than making  the ui transform a function and putting the text transform in the function making it change every time the button moves.