Textures for instanced rendering
i fucking hate the fragment shader i hate working on this. this took way to fucking long for what is was worth and it didn't even work the way i wanted but it works and that's all that matters. it runs better than sdl now. you can't fullscreen because entity limit is 1000 and it will crash but it works for now