BASIC NETWORKING A FUCKING GAIN!
boy howdy was this one sure a doozy! and guess what? we're using the harder version of the steam networking api!!! this might *technically* turn out to be easier since i can bug test faster but golly gee the steam api docs are the fucking forbidden archives that are so esoteric and give out no useful knowledge.